refactor for checkout to use fs::Context (#301)
The latter should be useful when fully implementing all required
baseline capabilities of doing a correct checkout
